correct bool(qApp) on Python 2
The idiom to create an application if none exists

    qApp or QApplication()

did not work on Python 2. This was recognized when another
unrelated test broke that tried to use the idiom.

This patch adds the missing method to Py_TYPE(Py_None)
which Python 3 already has.

A test case was added.

Side-effect.. The warning about missing braces is gone,
after the "PYTHON_IS_PYTHON3" macro is defined. This
was forgotten in the review when when the macro was
moved out of signature.cpp .

Task-number: PYSIDE-571
Change-Id: I9b9304f407cd4cafb3ba0945ffb045802503a343
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>