shiboken: fix a signature bug in __doc__ handling of global functions
The signature module was extended to allow automatic creation of
__doc__ attributes by taking str of the signature.
This function has to figure out when it should not touch an object
that does not come from PySide.

The check asked the type that was returned by GetClassOfFunc,
but in case of a global function, the module of the function is returned.

In order to prevent such mistakes, the function has been renamed
to GetClassOrModOf.

Change-Id: I055f1f0483674553f9300d642aa516cec607e66c
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Keith Kyzivat <keith.kyzivat@qt.io>