PySide: Create a framework for deprecated functions
During development of the patch
"Support Pointer Primitive Types by Arrays or Result Tuples"
some functions in QtGui turned out to be removal candidates.
The name "constData" should be deprecated in favor of the existing
"data" function. Other implementation also do not have this.

Instead of simply removing, we now create a surrogate function
with the name "constData" in fure Python that gives a warning
and calls the "data" function.

This is now extracted into its own commit since the deprecation
is a completely different and independent issue.
The implementation does not do any extra imports in advance.
and is easily extensible to more post-installation actions.

Task-number: PYSIDE-795
Change-Id: I410c69a87d9f0df78f736991b2ee0a2747678911
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>