Fix Xcode sdk value embedded into PySide2 binaries
Previously when building on the Coin VMs, the incorrect
Xcode sdk value of "10.0" was embedded into all the built shiboken
and PySide2 binaries. This causes issues when trying to notarize
a PySide2 app.

The reason why this happens is described as a long investigation
on the respective bug report.

The fix is to ask xcrun to report an SDK path which contains
the version number in the SDK path (/path/to/MacOSX10.14.sdk
vs just /path/to/MacOSX.sdk).

Change-Id: I3f02510953e2f54032c19a48cb8a7162814bf9e7
Fixes: PYSIDE-1066
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Simo Fält <simo.falt@qt.io>