Fix heaptype conflict with QtCore.QObject.__new__in Python 2.7
The patching of the type generation needs to become universal.
Additional to the patch for SbkObjectType_TypeF, we now patch

* SbkObjectTypeTpNew
* introduceWrapperType

which makes the modifications for the heaptype flag complete.

Actually, the modification of tp_new_wrapper has to be more
sophisticated in this general case:
We follow the __mro__ and patch exactly those functions which
have the old wrapper and add the new wrapper.

Change-Id: I51d4d77c99bd18eed2e31d2ab21143c0f4e2ea6c
Fixes: PYSIDE-1051
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>