Replace pyside2 uic/rcc
After the compatibility patches for uic and rcc,
is now possible to generate Python code by using the option
'-g python':
    uic -g python file.ui -o ui_file.py
    rcc -g python file.qrc -o rc_file.py

This patch depends on many changes on the 'pyside2-tools'
submodule, which mainly remove the old implementations of pyside2-uic
and pyside2-rcc.

The pyside_tool.py wrapper is now in charge of calling the uic and rcc
binaries, with the -g python every time the pyside2-uic and pyside2-rcc
binaries are called. To achieve this, we are now shipping the uic and rcc
binaries from the Qt installation.

Fixes: PYSIDE-1098
Change-Id: Ibdec0012f7ed671cd99424e1258c20649609c2da
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>