Enable the Function Registry for 5.14
The function registry was not enabled for versions greater than 5.12 .

This is now needed, since the function registry will be used in
the tests for the improved NumPy support.

There were new cases of Python keywords touched by enums
which had to be renamed. This was moved into the code
generator instead of the runtime overhead.

The formatting of the enums was rewritten and reports all
enums now that can be found (also those which are copied into
the parent class).

The formatting of the function registry had not
been used for a long time and had entries that showed the
wrong number of subclasses.

The usage of the registry was also simplified by using the full
names of functions. They can now directly be accessed.

Task-number: PYSIDE-795
Change-Id: I734f6811205f3c3528a911975677eb677fedd2dd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>