shiboken: Remove ShibokenGenerator::getDefaultValue()
Replace by AbstractMetaArgument::hasDefaultValueExpression(),
which is now restricted to the effective default value expression
after the previous commit.

This effectively reverts 96bd943d57b1864d1f5e7359b100c67d01934619 for
the old bug 919, which was trying to re-apply the modifications again,
potentially hitting on a remove modification in a loop.

Task-number: PYSIDE-1095
Change-Id: I08fc005cfdc2f7b91f8303abd4d75a479d670ba9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Tismer <tismer@stackless.com>