shiboken: Fix pyi generation for external modules
enum_sig.py is used to generate .pyi files (in
sources/pyside2/PySide2/support/generate_pyi.py) from the bindings.

The file can also be useful by external modules that were generated by
shiboken, however the check for the modulename being PySide2 or
shiboken2 prevents this.

Remove this check, as discussed on IRC. This seems to be working
internally as well and re-adds support for usages by external modules.

Change-Id: I81af943c43f8f84aa0b0d89260cd99215acdb10d
Reviewed-by: Christian Tismer <tismer@stackless.com>