Make the scrape script robust enough, again
The script scrape_testresults.py analyzes log files in order to
find embedded script files for the registry.
This script does not try to be absolute water-proof but handled the
existing cases well.

By chance, the text "BEGIN_FILE", which is used as an indicator,
was found in a traceback from the generating script.

We modified the search to ignore this text in single quotes, but
also changed the originating file testing/runner.py to no longer
contain this text in the future.

Change-Id: I1fef1cf5e9b85fc94b7ee52edbdf4afa4bf4e35a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>