Update the registry files after merging 5.13 into 5.14
The merge of certain changes into 5.14 had quite a few side effects,
because the registry files had not been generated for 5.13 at all.

When a simplification had the effect of generating more registry
entries, the merge broke because we did not have these entries in
the registry.

The "obvious" but wrong assumption was that the simplification
gives a problem in 5.14. But the truth is that by generating the
files before the merge created too few entries, and that made it
complain. A simple reset was the solution.

We will try to further simplify the generation of new registry files
without creating huge deltas.

Change-Id: I680b3a1d3eeb4b981736ff6fe46857ee3135b735
Reviewed-by: Christian Tismer <tismer@stackless.com>