Simplify recreation of registry files
The registry files were recreated when they do not exist. This was used
to trigger recreation on a file.

We now include a comment line on top of the files.

    #recreate       # uncomment this to enforce generation

By uncommenting this line, a NameError is provoked, which has the
desired effect without creating huge deltas in the repository.

Change-Id: Idcc1015abae504f111102e8c9851f8ef45dcbdff
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>