qApp: Ensure QtCore import when embedded QApplication subclass is used
The qApp machinery works great with Python.
When using embedding, things are different because there
is no longer a wrapper layer.

Unfortunately, many extension modules use C++ to derive a
QApplication class.

This has the side effect that when a foreign C++ module gets
imported, the qApp machinery does not see it as it would
in Python.

Instead of a complex analysis, we always make sure that QtCore
is imported. It will report the right instance, anyway.

This change could not easily be tested.
It was confirmed as a solution by Antonio Rojas.

Change-Id: Ie9c56ac75e6c0ae3ace615dfc26c6d218ff4efea
Fixes: PYSIDE-1135
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>