Replace OPTION_* by a dictionary
Currently we are importing * from the options.py file,
which is a problem for the linters to discover if we are properly
using the options of the file.

Having a dictionary provides also a better way of access these options,
and it is more clean than having one variable per each option.

Change-Id: Ie70bd88665357b85e2842c5df269857504ecccdf
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>