Brute solution for skipping missing ci template
There is a problem with Ubuntu 18.04 provisioning in
Qt5#5.12.6 branch, so we can't use that target. In
order to be able to skip that, we have to use CI's
custom module class instead of pyside class to get
pass the blocker. This commit is only to be used in
release branch, while it disables LicenceCheck target.

Change-Id: I5530d4ae19584685bbec034d0b8b977f25082ba2
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>