shiboken: Handle smart pointers with const pointees
Treat them equally to smart pointers with non-const
pointees by searching for the pointee type entry in the
instantiation search and stripping the const there. Create
an additional type index equal to the non-const pointee
index for the const pointee instantiation.

This has the slight drawback that it is possible to invoke
non-const calls on the pointee in Python, but this is nearly
unfixable since the dispatch happens in the getattro
method, where the information about constness is lost.

Task-number: PYSIDE-454
Change-Id: Ib514816e408d727656b1599060e105ccd021bb07
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>