Fix the registry after QTEST_ENVIRONMENT=ci was restored
Since the QTEST_ENVIRONMENT variable was no longer set,
a number of crucial checks were not performed.

One side effect are the two new keys which were missing in the registry.

Additionally, the registry is missing very many entries on macOS, and I
have no idea what the reason is. This could be an effect of the
macOS 10.15.1 version or of my maybe outdated Qt 5.14.0 version.
The registry files from 2019-10-31 have all entries. If I build
that version from 2019-10-31 and test it, I get the same missing keys.
Therefore, I doubt my results quite a bit!

To verify this, we simply check this change in, together with
https://codereview.qt-project.org/c/pyside/pyside-setup/+/284809

If that works, then I have a serious bug somewhere, but we know then
that the harm of that CI bug was a minimum. Fingers crossed!!

Change-Id: I25555d60d6911fca6de67110c35dff8d23c2fd8a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>