Fix Limited API for Python 3.8
This trivial patch fixes an omission that has not been caught
in earlier versions:

    PyObject_INIT

must be replaced by

    PyObject_Init

I think this change came by chance and was not the primary intent
of the authors. Otherwise they would have made sure that the
PyObject_INIT macro does not even exist when the Limited API
is enabled.

Change-Id: Iee8127dd398f1ec0997b9a629c5c48076137cecf
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>