examples: improve the star delegate example
While the example works correctly as is, if one takes the StarEditor
class alone, it will fail to work as it uses a starRating attribute that
is available in the example only  because of the call to setEditorData
by the StarDelegate class.

This patch fixes that as well makes use of the append overload of
QPolygonF taking a list of point to simplify the code a bit.

Change-Id: Id8132fa86f5668645f48c62377dac65f707c968b
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>