shiboken: Refactor writing of get/setattro methods
getattro functions are needed for:
  - smartpointer dispatch
  - special cases of function overload resolution

setattro functions are needed for:
  - smartpointer dispatch
  - QObject property handling

All of this was previously handled in the writeSet/GetattroFunction
which duplicated, inconsistent checks in various places, which makes
it very hard to add additional functionality to tp_getsetattro(). To
overcome this, define a flag to describe the various use cases, a
check function to determine it. The flag is passed to the write functions.
Smart pointer handling is split out completely since it will never mix
with the other use cases.

Task-number: PYSIDE-803
Change-Id: Iead9ee5b086830fb670b25917914117da62fefe5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>