Fix build on Windows using Python 3.8
The DLL load behavior was changed in Python 3.8 to no longer
search the PATH variable for DLL dependencies. This means
that the shiboken2.dll from the shiboken2 package is no longer found.

Add the shiboken2 package using os.add_dll_directory() to fix this.

Unfortunately, this requires additional work during the build process,
Qt and libpyside2.dll need to be found in the build directory.

Change-Id: I3d78d8df2d6f64913e06fa44e7e1aadbc1a14b58
Reviewed-by: Christian Tismer <tismer@stackless.com>