shiboken: Fix funny Python 2.7 glitches
There are two funny glitches that happen with Python 2.7, only.
One was the smartpointer test which only failed because
Python 2.7 has no list.clear method.
This test could be simplified since the tested object is a true list.

The other one appeared after the introduction of the BUILD_DIR
variable. Python 2.7 failed the enum_test.py because it now
needs an import of shiboken2 to do initialization of the signature
module. That is a timing problem that we did not have before
we had BUILD_DIR.

It makes me curious why this bug could slip through the COIN tests!

Change-Id: I9b509e9998ad9f40431a8b148892429c0e865393
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>