Fix qtinfo on macOS when XCode is newly installed
I happened to try out the "cling" C++ interpreter. That forced
me to install XCode. I did not run XCode because cling worked
immediately.

But then, building PySide always failed with an error in qtinfo.

This is a rare trap into which I fall every other year, so here is a
fix that returns the intended error message.

Change-Id: I9ec5774c49b8a186d6b4247150dac6b252e7cc64
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>