Accept EnumMeta as a valid Signal type
Including a check for Python EnumMeta types to not consider them
as a normal PySequence allows the decision to use the default
PyObject wrapper case.

Using the 'object' type is currently the workaround,
so this allow the users to use the Enum class instead of declaring
the signal with 'object'.

    class A(Enum):
        a = 1
        b = 1

    # Workaround
    # signal = Signal(object)

    # With this patch
    signal = Signal(A)

A test case was added.

Fixes: PYSIDE-239
Change-Id: Ib593dba5a988eceb8b1bfae097768e9ec02be6d5
Reviewed-by: Christian Tismer <tismer@stackless.com>