Avoid a signature warning in Python 3.6
Regular expressions which can evaluate to an empty string
result in a warning in Python 3.6 if you try to re.split() on
that pattern.

The generated pattern for the signature parser was enhanced
to avoid the unintended empty pattern.

This change was not very large, actually only a single bit
(but the right one). :-)

>>> bin(ord("*"))
'0b101010'
>>> bin(ord("+"))
'0b101011'

Change-Id: I15ba6ef6f108c51afba59b4004261bede26c4a74
Fixes: PYSIDE-1247
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>