shiboken: Update sbkstring to use PyUnicode_GetLength
The unicode implementation has changed very much.
PyUnicode_GET_SIZE has been used all the time, although this
macro is already a quite expensive real function since Python 3.3 .

The function is deprecated, and the macro PyUnicode_GET_LENGTH
should be used, instead.

This is relevant for cleaning things up, because in the course of
fixing PYSIDE-813 we work with debug Python, which then complains
that Python memory is used without holding the GIL.

The usage of the right implementation was ensured by a function
PepUnicode_GetLength and removing all traces of the former version.

Task-number: PYSIDE-813
Change-Id: I62e94e10e14975dac3dad0ed1fffec8a1b54a0d5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>