pep384impl.cpp: Make the implementation more stateless
The pep384impl.h part is quite readable and stateless concerning the
state of Py_LIMITED_API.
The pep384impl.cpp part is stateful, which means there is no clear
separation between parts of the implementation.

This had led to unnecessary confusion in the past. To avoid that in
future, the #ifdef status of Py_LIMITED_API will no longer be carried
over multiple sections.

Also, checks which are for the limited API only are no longer executed
by default.

Further, PyDict_GetItem is replaced by PyDict_GetItemWithError, where
we do not expect a real error.

Change-Id: Ia51e551216c76f82b701ebc45c40e2d1412cabf5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>